requirements.txt
backports.shutil-get-terminal-size==1.0.0
BeautifulSoup==3.2.1
decorator==4.0.9
DeliciousAPI==1.6.7
Django==1.9.5
djangorestframework==3.3.3
factory-boy==2.7.0
fake-factory==0.5.7
ipaddress==1.0.16
ipython==4.2.0
ipython-genutils==0.1.0
pathlib2==2.1.0
pexpect==4.0.1
pickleshare==0.7.2
ptyprocess==0.5.1
python-dateutil==2.5.3
simplegeneric==0.8.1
simplejson==3.8.2
six==1.10.0
traitlets==4.2.1
